学习抽象工厂和工厂方法
