+++ factory-method/src/main/java/com/iluwatar/factory/method/App.java
@@ -50,12 +50,14 @@ public class App {
    */
   public static void main(String[] args) {
 
+    // 兽人需要长矛
     Blacksmith blacksmith = new OrcBlacksmith();
     Weapon weapon = blacksmith.manufactureWeapon(WeaponType.SPEAR);
     LOGGER.info(MANUFACTURED, blacksmith, weapon);
     weapon = blacksmith.manufactureWeapon(WeaponType.AXE);
     LOGGER.info(MANUFACTURED, blacksmith, weapon);
 
+    // 精灵需要斧头
     blacksmith = new ElfBlacksmith();
     weapon = blacksmith.manufactureWeapon(WeaponType.SPEAR);
     LOGGER.info(MANUFACTURED, blacksmith, weapon);
